requirements.txt
tensorflow==2.11.0
Pillow==9.5.0